requirements.txt
tensorflow-gpu
smile
